@@ -1,39 +1,56 @@
 {
-  "name": "dolar-ui",
-  "version": "1.0.0",
-  "scripts": {
-    "ng": "ng",
-    "start": "ng serve",
-    "build": "ng build",
-    "watch": "ng build --watch --configuration development"
-  },
-  "private": true,
-  "dependencies": {
-    "@angular/animations": "~12.0.3",
-    "@angular/cdk": "^12.0.3",
-    "@angular/common": "~12.0.3",
-    "@angular/compiler": "~12.0.3",
-    "@angular/core": "~12.0.3",
-    "@angular/forms": "~12.0.3",
-    "@angular/platform-browser": "~12.0.3",
-    "@angular/platform-browser-dynamic": "~12.0.3",
-    "@angular/router": "~12.0.3",
-    "@ngx-translate/core": "^13.0.0",
-    "@ngx-translate/http-loader": "^6.0.0",
-    "luxon": "^2.0.2",
-    "primeflex": "^2.0.0",
-    "primeicons": "^4.1.0",
-    "primeng": "^12.2.1",
-    "rxjs": "~6.6.0",
-    "tslib": "^2.1.0",
-    "zone.js": "~0.11.4"
-  },
-  "devDependencies": {
-    "@angular-devkit/build-angular": "~12.0.3",
-    "@angular/cli": "~12.0.3",
-    "@angular/compiler-cli": "~12.0.3",
-    "@types/luxon": "^2.0.5",
-    "@types/node": "^12.11.1",
-    "typescript": "~4.2.3"
-  }
+    "name": "dolar-ui",
+    "version": "1.0.0",
+    "scripts": {
+        "ng": "ng",
+        "start": "ng serve",
+        "build": "ng build --configuration production",
+        "watch": "ng build --watch --configuration development",
+        "server": "node server.js",
+        "prettier": "prettier ./src --write",
+        "eslint": "eslint ./src --max-warnings 0 --no-warn-ignored",
+        "prepare": "husky"
+    },
+    "private": true,
+    "dependencies": {
+        "@angular/animations": "^17.3.10",
+        "@angular/cdk": "^17.3.10",
+        "@angular/common": "^17.3.10",
+        "@angular/compiler": "^17.3.10",
+        "@angular/core": "^17.3.10",
+        "@angular/forms": "^17.3.10",
+        "@angular/platform-browser": "^17.3.10",
+        "@angular/platform-browser-dynamic": "^17.3.10",
+        "@angular/router": "^17.3.10",
+        "@ngx-translate/core": "^15.0.0",
+        "@ngx-translate/http-loader": "^8.0.0",
+        "luxon": "^3.4.4",
+        "primeflex": "^3.3.1",
+        "primeicons": "^7.0.0",
+        "primeng": "^17.18.0",
+        "rxjs": "~6.6.0",
+        "tslib": "^2.1.0",
+        "zone.js": "~0.14.6"
+    },
+    "devDependencies": {
+        "@angular-devkit/build-angular": "^17.3.8",
+        "@angular/cli": "^17.3.8",
+        "@angular/compiler-cli": "^17.3.10",
+        "@eslint/js": "^8.57.0",
+        "@types/luxon": "^3.4.2",
+        "@types/node": "^18.19.33",
+        "eslint": "^8.57.0",
+        "globals": "^15.6.0",
+        "husky": "^9.0.11",
+        "lint-staged": "^15.2.5",
+        "prettier": "^3.3.0",
+        "typescript": "~5.4.5",
+        "typescript-eslint": "^7.13.1"
+    },
+    "lint-staged": {
+        "**/*": [
+            "prettier --write --ignore-unknown",
+            "eslint --max-warnings 0 --no-warn-ignored"
+        ]
+    }
 }
