@@ -1,8 +1,8 @@
 import { animate, style, transition, trigger } from '@angular/animations';
 
 export const slideDownAndFadeIn = trigger('slideDownAndFadeIn', [
-  transition('void => *', [
-    style({ opacity: 0, transform: 'translateY(-25px)' }),
-    animate('0.5s')
-  ])
+    transition('void => *', [
+        style({ opacity: 0, transform: 'translateY(-25px)' }),
+        animate('0.5s'),
+    ]),
 ]);
