@@ -4,16 +4,16 @@ import { PrimeNGConfig } from 'primeng/api';
 import { slideDownAndFadeIn } from './animations/animations';
 
 @Component({
-  selector: 'app-root',
-  templateUrl: './app.component.html',
-  styleUrls: ['./app.component.css'],
-  animations: [
-    slideDownAndFadeIn
-  ]
+    selector: 'app-root',
+    templateUrl: './app.component.html',
+    styleUrls: ['./app.component.scss'],
+    animations: [slideDownAndFadeIn],
 })
 export class AppComponent {
-  
-  constructor(private config: PrimeNGConfig, private translateService: TranslateService) {}
+    constructor(
+        private config: PrimeNGConfig,
+        private translateService: TranslateService
+    ) {}
 
     ngOnInit() {
         const lingua = 'pt';
@@ -22,11 +22,13 @@ export class AppComponent {
     }
 
     /**
-     * Faz a tradução do componente do primeng (Calendário) para o português 
+     * Faz a tradução do componente do primeng (Calendário) para o português
      * @param lang Nome da língua
      */
     traduzir(lang: string) {
         this.translateService.use(lang);
-        this.translateService.get('primeng').subscribe(res => this.config.setTranslation(res));
+        this.translateService
+            .get('primeng')
+            .subscribe((res) => this.config.setTranslation(res));
     }
 }
