@@ -1,11 +1,11 @@
 export interface CotacaoDolar {
-  valor: number;
-  data: Date;
-  diferenca: number;
+    valor: number;
+    data: Date;
+    diferenca: number;
 }
 
 export interface CotacaoDolarDTO {
-  valor: string;
-  data: string;
-  diferenca: string;
+    valor: string;
+    data: string;
+    diferenca: string;
 }
