@@ -1,28 +1,28 @@
 export interface Page<T> {
-  content: Array<T>;
-  pageable: Pageable;
-  last: boolean;
-  totalPages: number;
-  totalElements: number;
-  number: number;
-  size: number;
-  sort: Sort;
-  first: boolean;
-  numberOfElements: number;
-  empty: boolean;
+    content: Array<T>;
+    pageable: Pageable;
+    last: boolean;
+    totalPages: number;
+    totalElements: number;
+    number: number;
+    size: number;
+    sort: Sort;
+    first: boolean;
+    numberOfElements: number;
+    empty: boolean;
 }
 
 interface Pageable {
-  sort: Sort;
-  offset: number;
-  pageNumber: number;
-  pageSize: number;
-  unpaged: boolean;
-  paged: boolean;
+    sort: Sort;
+    offset: number;
+    pageNumber: number;
+    pageSize: number;
+    unpaged: boolean;
+    paged: boolean;
 }
 
 interface Sort {
-  empty: boolean;
-  sorted: boolean;
-  unsorted: boolean;
+    empty: boolean;
+    sorted: boolean;
+    unsorted: boolean;
 }
