@@ -1,4 +1,4 @@
 export const environment = {
-  production: true,
-  apiUrl: ''
+    production: true,
+    apiUrl: 'http://localhost:8080',
 };
