@@ -3,8 +3,8 @@
 // The list of file replacements can be found in `angular.json`.
 
 export const environment = {
-  production: false,
-  apiUrl: 'http://localhost:8080'
+    production: false,
+    apiUrl: 'http://localhost:8080',
 };
 
 /*
